requirements.txt
langchain-core~=0.2.22
langchain-openai~=0.1.17
python-dotenv~=1.0.1
fastapi~=0.111.1
langserve~=0.2.2
uvicorn~=0.31.0